+++ saul-examples/src/main/scala/edu/illinois/cs/cogcomp/saulexamples/mSpRL2017/Helpers/DataProportion.scala
@@ -1,0 +1,9 @@
+package edu.illinois.cs.cogcomp.saulexamples.mSpRL2017.Helpers
+
+/**
+  * Created by taher on 2017-02-28.
+  */
+object DataProportion extends Enumeration {
+  type DataProportion = Value
+  val Train, ValidationTrain, ValidationTest, Test, All = Value
+}
