+++ saul-examples/src/main/scala/edu/illinois/cs/cogcomp/saulexamples/mSpRL2017/Helpers/FeatureSets.scala
@@ -1,0 +1,9 @@
+package edu.illinois.cs.cogcomp.saulexamples.mSpRL2017.Helpers
+
+/**
+  * Created by taher on 2017-02-28.
+  */
+object FeatureSets extends Enumeration {
+  type FeatureSets = Value
+  val BaseLine, WordEmbedding, WordEmbeddingPlusImage = Value
+}
