+++ saul-examples/src/main/scala/edu/illinois/cs/cogcomp/saulexamples/mSpRL2017/MultiModalConstrainedClassifiers.scala
@@ -2,11 +2,10 @@ package edu.illinois.cs.cogcomp.saulexamples.mSpRL2017
 
 import edu.illinois.cs.cogcomp.infer.ilp.OJalgoHook
 import edu.illinois.cs.cogcomp.saul.classifier.ConstrainedClassifier
-import edu.illinois.cs.cogcomp.saulexamples.nlp.BaseTypes.{Relation, Sentence, Token}
-import MultiModalSpRLConstraints._
-import MultiModalSpRLDataModel._
 import edu.illinois.cs.cogcomp.saulexamples.mSpRL2017.MultiModalSpRLClassifiers._
 import edu.illinois.cs.cogcomp.saulexamples.mSpRL2017.MultiModalSpRLConstraints._
+import edu.illinois.cs.cogcomp.saulexamples.mSpRL2017.MultiModalSpRLDataModel.{relationToFirstArgument, relationToSecondArgument}
+import edu.illinois.cs.cogcomp.saulexamples.nlp.BaseTypes.{Relation, Token}
 /** Created by parisakordjamshidi on 2/4/17.
   */
 object MultiModalConstrainedClassifiers {
