+++ saul-examples/src/main/scala/edu/illinois/cs/cogcomp/saulexamples/mSpRL2017/SentenceLevelConstraintClassifiers.scala
@@ -3,9 +3,9 @@ package edu.illinois.cs.cogcomp.saulexamples.mSpRL2017
 import edu.illinois.cs.cogcomp.infer.ilp.OJalgoHook
 import edu.illinois.cs.cogcomp.saul.classifier.ConstrainedClassifier
 import edu.illinois.cs.cogcomp.saulexamples.mSpRL2017.MultiModalSpRLClassifiers._
+import edu.illinois.cs.cogcomp.saulexamples.mSpRL2017.MultiModalSpRLDataModel.{sentenceToRelations, sentenceToToken}
 import edu.illinois.cs.cogcomp.saulexamples.mSpRL2017.SentenceLevelConstraints._
 import edu.illinois.cs.cogcomp.saulexamples.nlp.BaseTypes.{Relation, Sentence, Token}
-import MultiModalSpRLDataModel.{sentences, _}
 
 /** Created by parisakordjamshidi on 2/9/17.
   */
