+++ saul-examples/src/main/scala/edu/illinois/cs/cogcomp/saulexamples/mSpRL2017/SpRLApp.scala
@@ -4,13 +4,13 @@ import java.io.File
 
 import edu.illinois.cs.cogcomp.core.utilities.XmlModel
 import edu.illinois.cs.cogcomp.saul.util.Logging
+import edu.illinois.cs.cogcomp.saulexamples.mSpRL2017.Helpers.FeatureSets
 import edu.illinois.cs.cogcomp.saulexamples.mSpRL2017.MultiModalSpRLClassifiers._
 import edu.illinois.cs.cogcomp.saulexamples.nlp.BaseTypes.{Document, NlpBaseElement, Sentence, Token}
 import edu.illinois.cs.cogcomp.saulexamples.nlp.SpatialRoleLabeling.SpRL2013.{LANDMARK, RELATION, SPATIALINDICATOR, TRAJECTOR}
 import edu.illinois.cs.cogcomp.saulexamples.nlp.SpatialRoleLabeling._
 
 import scala.collection.JavaConversions._
-import scala.collection.mutable
 
 /**
   * Created by taher on 2017-02-24.
@@ -61,7 +61,7 @@ object SpRLApp extends App with Logging {
       new Document(file.getName, -1, -1, text)
     }).toList
 
-    MultiModalPopulateData.populateDataFromRawDocuments(documentList)
+    MultiModalPopulateData.populateDataFromPlainTextDocuments(documentList)
 
     val relationList = TripletClassifierUtils.predict(
       x => TrajectorPairClassifier(x),
