+++ saul-examples/src/main/scala/edu/illinois/cs/cogcomp/saulexamples/mSpRL2017/TextSceneCrossConstraints.scala
@@ -1,10 +1,4 @@
 package edu.illinois.cs.cogcomp.saulexamples.mSpRL2017
-import edu.illinois.cs.cogcomp.lbjava.infer.{FirstOrderConstant, FirstOrderConstraint}
-import edu.illinois.cs.cogcomp.saul.classifier.ConstrainedClassifier
-import edu.illinois.cs.cogcomp.saulexamples.mSpRL2017.MultiModalSpRLClassifiers._
-import edu.illinois.cs.cogcomp.saulexamples.mSpRL2017.MultiModalSpRLDataModel._
-import edu.illinois.cs.cogcomp.saulexamples.nlp.BaseTypes.{Relation, Sentence, Token}
-import edu.illinois.cs.cogcomp.saul.constraint.ConstraintTypeConversion._
 
 /**
   * Created by parisakordjamshidi on 2/14/17.
